pyproject.toml
[tool.poetry]
name = "valveprotos_py"
version = "0.1.0"
description = ""
authors = ["Manuel Raimann <raimannma@outlook.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
protobuf = "^5.29.3"

[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
PyGithub = "^2.5.0"
protobuf-to-pydantic = {version = "^0.3.0.3", extras = ["mypy-protobuf"]}
toml = "^0.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
